requirements.txt
numpy
matplotlib
torch
tensorflow
transformers
scikit-learn
liblinear-official
gensim